Replaced core of notebook object.
