+++ .gitignore
@@ -124,3 +124,4 @@ dmypy.json
 
 # VS Code
 *.code-workspace
+.vscode/
