+++ src/test.py
@@ -1,0 +1,5 @@
+import tidynotes
+
+# test = tidynotes.Notebook.initialise(r"C:\Users\Joe\Documents\My_Temp\test")
+test = tidynotes.Notebook(r"C:\Users\Joe\Documents\My_Temp\notes")
+test.clean()
