+++ src/tidynotes/__init__.py
@@ -2,7 +2,7 @@
 Tools to generate and manage Markdown notebooks.
 """
 
-from .notebook import Tidybook
+from .notebook import Notebook
 
 # Version of the tidynotes package
 __version__ = "20.05.01"
