+++ src/tidynotes/templates/note.md
@@ -1,1 +1,1 @@
-# {{dates['title']}}
+# {{date.strftime('%Y-%m-%d (%A)')}}
